pyproject.toml
[project]
name = "fb_viewer"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "a170811", email = "a170811@gmail.com"},
]
dependencies = [
    "beautifulsoup4>=4.12.3",
    "python-dotenv>=1.0.1",
    "selenium>=4.24.0",
    "loguru>=0.7.2",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "ipython>=8.27.0",
    "lxml>=5.3.0",
]
